+++ composeApp/src/commonMain/kotlin/com/virtualworld/mipymeanabelmaster/core/dto/Product.kt
@@ -1,4 +1,4 @@
-package com.virtualworld.mipymeanabelmaster.core
+package com.virtualworld.mipymeanabelmaster.core.dto
 
 import kotlinx.serialization.Serializable
