+++ composeApp/src/commonMain/kotlin/com/virtualworld/mipymeanabelmaster/screen/orders/OrdersViewModel.kt
@@ -2,8 +2,7 @@ package com.virtualworld.mipymeanabelmaster.screen.orders
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.virtualworld.mipymeanabelmaster.core.NetworkResponseState
-import com.virtualworld.mipymeanabelmaster.core.Product
+import com.virtualworld.mipymeanabelmaster.core.model.NetworkResponseState
 import com.virtualworld.mipymeanabelmaster.core.dto.Order
 import com.virtualworld.mipymeanabelmaster.domain.GetOrdersSentUseCase
 import kotlinx.coroutines.flow.MutableStateFlow
@@ -13,7 +12,8 @@ import kotlinx.coroutines.launch
 
 class OrdersViewModel(private val getOrdersSent: GetOrdersSentUseCase) : ViewModel() {
 
-    private val _ordersSent = MutableStateFlow<NetworkResponseState<List<Order>>>(NetworkResponseState.Loading)
+    private val _ordersSent = MutableStateFlow<NetworkResponseState<List<Order>>>(
+        NetworkResponseState.Loading)
     val ordersSent: StateFlow<NetworkResponseState<List<Order>>> = _ordersSent
 
     init {
